Replace the tokenManager denylist with a real allowlist module
A denylist (importNames: TOKEN_MANAGER_DATA_ACCESS_EXPORTS) silently
permits any future tokenManager.ts export by default — exactly the
maintenance risk flagged in review. ESLint's no-restricted-imports has
no "allow only these names" mode, so fix this structurally instead of
in config: move the six pure OAuth CSRF state helpers (no token/network
access) into their own module, src/auth/oauthState.ts.

The UI-layer ESLint block now blocks auth/tokenManager outright for
pages/components/layouts/contexts, with no per-file exception — the
three documented pages (and useGitHubAuth.ts, ProtectedRoute.tsx,
AuthContext.tsx) import the CSRF helpers from oauthState.ts instead,
which the rule doesn't restrict. A new tokenManager.ts export is now
blocked from UI code by default; it would have to be added to
oauthState.ts first to become reachable there. Updated AGENTS.md and
src/pages/AGENTS.md to describe the new module instead of the old
exception list.

Verified: a page importing a never-seen-before tokenManager export
(simulating a future addition) is now rejected; the three real pages
still pass with their actual oauthState imports; the full lint, the
81-test unit suite, and a scoped tsc check of every touched file are
all clean.
