+++ ipaas/src/auth/ProtectedRoute.tsx
@@ -23,7 +23,7 @@ import { useAuth } from './AuthContext';
 import { useAccessControl } from '../contexts/AccessControlContext';
 import { fetchOrgPermissions } from '#api/auth';
 import { loginUrl, forceChangePasswordUrl } from '../paths';
-import { saveRedirectUrl } from './tokenManager';
+import { saveRedirectUrl } from './oauthState';
 import { Permissions } from '../constants/permissions';
 
 export default function ProtectedRoute(): JSX.Element {
