+++ ipaas/src/auth/tokenManager.ts
@@ -23,8 +23,6 @@ const ACCESS_TOKEN_KEY = 'icp_auth_token';
 const REFRESH_TOKEN_KEY = 'icp_refresh_token';
 const TOKEN_EXPIRES_AT_KEY = 'icp_token_expires_at';
 const REFRESH_TOKEN_EXPIRES_AT_KEY = 'icp_refresh_token_expires_at';
-const REDIRECT_URL_KEY = 'icp_redirect_url';
-const OIDC_STATE_KEY = 'icp_oidc_state';
 const OIDC_AUTH_MODE_KEY = 'icp_auth_mode';
 const OIDC_ORG_HANDLE_KEY = 'icp_org_handle';
 
@@ -413,43 +411,6 @@ export async function revokeToken(): Promise<void> {
   } catch {
     // best-effort — ignore errors
   }
-}
-
-export function saveRedirectUrl(url: string): void {
-  localStorage.setItem(REDIRECT_URL_KEY, url);
-}
-
-export function getAndClearRedirectUrl(): string | null {
-  const url = localStorage.getItem(REDIRECT_URL_KEY);
-  localStorage.removeItem(REDIRECT_URL_KEY);
-  return url;
-}
-
-export function generateAndSaveOIDCState(): string {
-  const state = crypto.randomUUID();
-  localStorage.setItem(OIDC_STATE_KEY, state);
-  return state;
-}
-
-export function validateAndClearOIDCState(state: string): boolean {
-  const savedState = localStorage.getItem(OIDC_STATE_KEY);
-  localStorage.removeItem(OIDC_STATE_KEY);
-  return savedState === state;
-}
-
-// GitHub OAuth CSRF state — sessionStorage so it's scoped to the initiating tab
-const GITHUB_OAUTH_STATE_KEY = 'icp_github_oauth_state';
-
-export function generateAndSaveGitHubState(): string {
-  const state = crypto.randomUUID();
-  sessionStorage.setItem(GITHUB_OAUTH_STATE_KEY, state);
-  return state;
-}
-
-export function validateAndClearGitHubState(state: string): boolean {
-  const saved = sessionStorage.getItem(GITHUB_OAUTH_STATE_KEY);
-  sessionStorage.removeItem(GITHUB_OAUTH_STATE_KEY);
-  return saved !== null && saved === state;
 }
 
 // ---------------------------------------------------------------------------
